fix: 스크롤 잠금 관련 에러 해결
