+++ src/lib/hooks/common/useLockScroll.tsx
@@ -1,9 +1,6 @@
-import { usePathname } from "next/navigation";
 import { useEffect } from "react";
 
 const useLockScroll = (isOpen?: boolean) => {
-  const pathname = usePathname();
-
   useEffect(() => {
     if (isOpen === false) return;
 
@@ -17,12 +14,6 @@ const useLockScroll = (isOpen?: boolean) => {
       document.documentElement.style.paddingRight = "";
     };
   }, [isOpen]);
-
-  // pathname이 변경되면 스크롤을 다시 활성화
-  useEffect(() => {
-    document.documentElement.style.overflow = "";
-    document.documentElement.style.paddingRight = "";
-  }, [pathname]);
 };
 
 export default useLockScroll;
